restructure project by organizing files into packages for better readability
