--- src/main/kotlin/Main.kt
@@ -1,9 +1,0 @@
-package org.example
-
-import org.example.data.CSVMealsRepository
-import java.io.File
-
-fun main() {
-    val meals  = CSVMealsRepository(File("food.csv")).getAllMeals()
-
-}
