+++ src/main/kotlin/data/CSVFoodRepository.kt
@@ -1,14 +1,15 @@
 package org.example.data
 
+import logic.models.Meal
 import org.example.logic.MealsRepository
-import org.example.models.Meal
 import java.io.File
 
-class CSVMealsRepository(private val csvFile: File): MealsRepository {
-    override fun getAllMeals(): List<Meal>{
+class CSVMealsRepository(private val csvFile: File) : MealsRepository {
+    override fun getAllMeals(): List<Meal> {
         return emptyList()
     }
-    private fun parseOneLine(line: String){
+
+    private fun parseOneLine(line: String) {
 
     }
 }
