+++ src/main/kotlin/logic/MealsRepository.kt
@@ -1,6 +1,6 @@
 package org.example.logic
 
-import org.example.models.Meal
+import logic.models.Meal
 
 interface MealsRepository {
     fun getAllMeals(): List<Meal>
