+++ src/main/kotlin/logic/models/Meal.kt
@@ -1,4 +1,4 @@
-package org.example.models
+package logic.models
 
 import java.time.LocalDate
