+++ src/main/kotlin/presentation/Main.kt
@@ -1,0 +1,9 @@
+package org.example.presentation
+
+import org.example.data.CSVMealsRepository
+import java.io.File
+
+fun main() {
+    val meals = CSVMealsRepository(File("food.csv")).getAllMeals()
+
+}
